Flutter driver commands refactoring
