+++ metrics/test_driver/common/command/command.dart
@@ -1,0 +1,12 @@
+library command;
+
+import '../config/browser_name.dart';
+import '../config/device.dart';
+
+part 'drive_command.dart';
+
+part 'flutter_command.dart';
+
+part 'run_command.dart';
+
+part 'run_command_base.dart';
