+++ metrics/test_driver/common/command/drive_command.dart
@@ -1,134 +1,13 @@
-import '../config/browser_name.dart';
-import '../config/device.dart';
+part of command;
 
 /// Wrapper class for the 'flutter drive' command.
 ///
 /// https://github.com/flutter/flutter/blob/f0a175f41d67c28312025a78278d4498ee488e4e/packages/flutter_tools/lib/src/commands/drive.dart#L116
 /// flutter drive --help
-class DriveCommand {
-  final List<String> _args = [];
-
+class DriveCommand extends _RunCommandBase {
   DriveCommand() {
     _add("drive");
   }
-
-  /// Method to create the [List] of args from the [DriveCommand] class instance
-  List<String> buildArgs() => List.unmodifiable(_args);
-
-  /// --device
-  ///
-  /// Target device id or name (prefixes allowed).
-  void device(Device device) => _add('--device-id=${device.deviceId}');
-
-  /// --verbose
-  ///
-  /// Noisy logging, including all shell commands executed.
-  void verbose() => _add('--verbose');
-
-  ///  --debug
-  ///
-  ///  Build a debug version of your app (default mode).
-  void debug() => _add("--debug");
-
-  ///  --profile
-  ///
-  ///  Build a version of your app specialized for performance profiling.
-  void profile() => _add("--profile");
-
-  ///  --release
-  ///
-  ///  Build a release version of your app.
-  void release() => _add("--release");
-
-  ///  --flavor
-  ///
-  ///  Build a custom app flavor as defined by platform-specific build setup.
-  ///  Supports the use of product flavors in Android Gradle scripts, and the
-  ///  use of custom Xcode schemes.
-  void flavor() => _add("--flavor");
-
-  ///  --trace-startup
-  ///
-  ///  Trace application startup, then exit, saving the trace to a file.
-  void traceStartup() => _add("--trace-startup");
-
-  ///  --verbose-system-logs
-  ///
-  ///  Include verbose logging from the flutter engine.
-  void verboseSystemLogs() => _add("--verbose-system-logs");
-
-  ///  --cache-sksl
-  ///
-  ///  Only cache the shader in SkSL instead of binary or GLSL.
-  void cacheSksl() => _add("--cache-sksl");
-
-  ///  --dump-skp-on-shader-compilation
-  ///
-  ///  Automatically dump the skp that triggers new shader compilations.
-  ///  This is useful for wrting custom ShaderWarmUp to reduce jank. By
-  ///  default, this is not enabled to reduce the overhead. This is only
-  ///  available in profile or debug build.
-  void dumpSkpOnShaderCompilation() => _add("--dump-skp-on-shader-compilation");
-
-  ///  --route
-  ///
-  ///  Which route to load when running the app.
-  void route(String route) => _add("--route=$route");
-
-  ///  --vmservice-out-file=<project/example/out.txt>
-  ///
-  ///  A file to write the attached vmservice uri to after an
-  ///  application is started.
-  void vmserviceOutFile(String file) => _add("--vmservice-out-file=$file");
-
-  ///  -t, --target=<path>
-  ///
-  ///  The main entry-point file of the application, as run on the device.
-  ///  If the --target option is omitted, but a file name is provided on the
-  ///  command line, then that is used instead.
-  ///  (defaults to "lib/main.dart")
-  void target(String path) => _add("--target=$path");
-
-  ///  --device-vmservice-port
-  ///
-  ///  Look for vmservice connections only from the specified port.
-  ///  Specifying port 0 (the default) will accept the first vmservice discovered.
-  void deviceVmservicePort(int port) => _add("--device-vmservice-port=$port");
-
-  ///  --host-vmservice-port
-  ///
-  ///  When a device-side vmservice port is forwarded to a host-side port,
-  ///  use this value as the host port.
-  ///  Specifying port 0 (the default) will find a random free host port.
-  void hostVmservicePort(int port) => _add("--host-vmservice-port=$port");
-
-  ///  --[no-]pub
-  ///
-  ///  Whether to run "flutter pub get" before executing this command.
-  ///  (defaults to on)
-  void pub() => _add("--pub");
-
-  ///  --[no-]pub
-  ///
-  ///  Whether to run "flutter pub get" before executing this command.
-  ///  (defaults to on)
-  void noPub() => _add("--no-pub");
-
-  ///  --[no-]track-widget-creation
-  ///
-  ///  Track widget creation locations. This enables features such as the
-  ///  widget inspector. This parameter is only functional in debug mode (i.e.
-  ///  when compiling JIT, not AOT).
-  ///  (defaults to on)
-  void trackWidgetCreation() => _add("--track-widget-creation");
-
-  ///  --[no-]track-widget-creation
-  ///
-  ///  Track widget creation locations. This enables features such as the
-  ///  widget inspector. This parameter is only functional in debug mode (i.e.
-  ///  when compiling JIT, not AOT).
-  ///  (defaults to on)
-  void noTrackWidgetCreation() => _add("--no-track-widget-creation");
 
   ///  --[no-]keep-app-running
   ///
@@ -217,7 +96,4 @@ class DriveCommand {
   void browserDimension(String browserDimension) {
     _add("--browser-dimension=$browserDimension");
   }
-
-  /// Private method to add the arg to [_args] list
-  void _add(String value) => _args.add(value);
 }
